give each class its own file in proper module
