+++ motion-cocoapods.gemspec
@@ -5,11 +5,11 @@ end; end; end
 
 require 'date'
 $:.unshift File.expand_path('../lib', __FILE__)
-require 'motion/project/version'
+require 'motion/cocoapods/version'
 
 Gem::Specification.new do |spec|
   spec.name        = 'motion-cocoapods'
-  spec.version     = Motion::Project::CocoaPods::VERSION
+  spec.version     = Motion::CocoaPods::VERSION
   spec.date        = Date.today
   spec.summary     = 'CocoaPods integration for RubyMotion projects'
   spec.description = "motion-cocoapods allows RubyMotion projects to have access to the CocoaPods dependency manager."
