+++ spec/spec_helper.rb
@@ -37,4 +37,3 @@ module SpecHelper
     end
   end
 end
-
